pyproject.toml
[build-system]
requires = [
    'setuptools ~= 68.2.2',
    'wheel ~= 0.41.2'
]
build-backend = 'setuptools.build_meta'

[project]
name = "mrmat-xmas-2023"
description = "Christmas Card 2023"
urls = { "Sources" = "https://github.com/MrMatAP/mrmat-xmas-2023" }
keywords = ["experimental"]
readme = "README.md"
license = { text = "MIT" }
authors = [
    { "name" = "Mathieu Imfeld", "email" = "imfeldma+9jqerw@gmail.com" }
]
maintainers = [
    { "name" = "Mathieu Imfeld", "email" = "imfeldma+9jqerw@gmail.com" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT",
    "Programming Language :: Python :: 3.11"
]
requires-python = ">=3.10"
dependencies = [
    "fastapi~=0.103.1",          # MIT
    "uvicorn~=0.23.2",           # BSD 3-Clause
    "httpx~=0.25.0",             # BSD 3-Clause
    "msal~=1.24.1",              # MIT
    "identity~=0.3.0",           # MIT
    "pyyaml~=6.0.1",             # MIT
    'fastapi-azure-auth~=4.2.1', # MIT
    'azure-cosmos~=4.5.1',       # MIT
    'azure-identity~=1.4.0',     # MIT
]
dynamic = ["version"]

[tool.setuptools.dynamic]
version = { attr = "ci.version "}

[tool.setuptools.packages.find]
where = ["src"]
include = ["mrmat_xmas_2023*"]
namespaces = true

[tool.setuptools.package-data]
"*" = [".mo", "*.html", "*.png", "*.webmanifest", "*.ico", "*.css", "*.js"]

# If you are debugging your tests using PyCharm then comment out the coverage options
# in addopts
[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--cov=mrmat_xmas_2023 --cov-report=term --cov-report=xml:build/coverage.xml --junit-xml=build/junit.xml"
testpaths = ["tests"]
junit_family = "xunit2"
log_cli = 1
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format="%Y-%m-%d %H:%M:%S"
